+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/FilterTableModel.java
@@ -31,6 +31,7 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
@@ -194,6 +195,14 @@ public interface FilterTableModel<R, C> extends TableModel, FilterModel<R> {
 		 * @return this builder instance
 		 */
 		Builder<R, C> async(boolean async);
+
+		/**
+		 * By default, exceptions during refresh are rethrown,
+		 * use this method to handle async exceptions differently
+		 * @param onRefreshException the exception handler to use during refresh
+		 * @return this builder instance
+		 */
+		Builder<R, C> onRefreshException(Consumer<Exception> onRefreshException);
 
 		/**
 		 * @param editor supplies the row editor