+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditComponentPanel.java
@@ -523,7 +523,6 @@ public class EntityEditComponentPanel extends JPanel {
 	 */
 	protected final <T, C extends JComboBox<T>, B extends ComboBoxBuilder<C, T, B>> ComboBoxBuilder<C, T, B> createComboBox(Column<T> column) {
 		FilterComboBoxModel<T> comboBoxModel = editModel().comboBoxModel(column);
-		comboBoxModel.items().refresher().exception().addConsumer(this::onException);
 
 		return (ComboBoxBuilder<C, T, B>) setComponentBuilder(column, entityComponents.comboBox(column, comboBoxModel)
 						.onSetVisible(EntityEditComponentPanel::refreshIfCleared));
@@ -536,7 +535,6 @@ public class EntityEditComponentPanel extends JPanel {
 	 */
 	protected final EntityComboBox.Builder createComboBox(ForeignKey foreignKey) {
 		EntityComboBoxModel comboBoxModel = editModel().comboBoxModel(foreignKey);
-		comboBoxModel.items().refresher().exception().addConsumer(this::onException);
 
 		return setComponentBuilder(foreignKey, entityComponents.comboBox(foreignKey, comboBoxModel)
 						.onSetVisible(EntityEditComponentPanel::refreshIfCleared));
@@ -551,7 +549,6 @@ public class EntityEditComponentPanel extends JPanel {
 	protected final EntityComboBoxPanel.Builder createComboBoxPanel(ForeignKey foreignKey,
 																																	Supplier<EntityEditPanel> editPanel) {
 		EntityComboBoxModel comboBoxModel = editModel().comboBoxModel(foreignKey);
-		comboBoxModel.items().refresher().exception().addConsumer(this::onException);
 
 		return setComponentBuilder(foreignKey, entityComponents.comboBoxPanel(foreignKey, comboBoxModel, editPanel))
 						.onSetVisible(entityComboBoxPanel -> refreshIfCleared(entityComboBoxPanel.comboBox()));