+++ changelog.md
@@ -8,8 +8,11 @@ Codion Change Log
 ### is.codion.common.rmi
 - SerializationFilterDryRun no longer includes primitives in output.
 - SerializationFilterDryRun now appends exclude all to output.
+### is.codion.common.model
+- FilterModel.Refresher.exception() observer removed, related refactoring.
 ### is.codion.swing.common.model
 - DelayedAction.Performer now ActionListener instead of Action.
+- AbstractRefreshWorker onRefreshException handler parameter added.
 ### is.codion.swing.common.ui
 - LookAndFeelSelectionDialogBuilder.allowInstalled() added along with ALLOW_INSTALLED configuration property.
 - ImagePanel can now have multiple overlays.